->superadmin listede olmayacak ->Toplu ekelem sayfası bilgilendirme tarih ayarı ->Alacak İzin double yapılacak ->Toplu yükleme butonu yükleniyor ibaresi eklendi ->Personel listesi dogum tarihi eklendi ->İban dogru düzende olması lazım ->Dashboard bekleyen izin oluşturulma tarihi azalana değiştirildi ->Kullanıcı Düzenleme username ->>> ad soyad ->Dashboard Gıda yardım maliyeti eklendi ->Telefon dogru düzende olması lazım ->table responsive ve no-wrap eklenecek tablolar ekran kuculunce ekrandan tasmıyor yana kaydırmalı ->Çıkarılan personel default sıralama çıkış tarihine göre azalan ->Detaylı kümülatif 2003 öncesi durumu eklenecek

+++ Core/DTOs/OffDayDTOs/ReadDtos/ReadWaitingOffDayEditDto.cs
@@ -37,5 +37,5 @@ public class ReadWaitingOffDayEditSubPersonalDto
     public string NameSurname { get; set; }
     public int TotalYearLeave { get; set; }
     public int UsedYearLeave { get; set; }
-    public int TotalTakenLeave { get; set; }
+    public double TotalTakenLeave { get; set; }
 }
